Ghostbusters Ghost Problem
A group of n Ghostbusters is battling n ghosts. Each Ghostbuster carries
a proton pack, which shoots a stream at a ghost, eradicating it. A
stream goes in a straight line and terminates when it hits the ghost.
The Ghostbusters decide upon the following strategy. They will pair off
with the ghosts, forming n Ghostbuster-ghost pairs, and then
simultaneously each Ghostbuster will shoot a stream at his chosen ghost.
As we all know, it is very dangerous to let streams cross, and so the
Ghostbusters must choose pairings for which no streams will cross.
Assume that the position of each Ghostbuster and each ghost is a fixed
point in the plane and that no three positions are collinear.